Boolevator: pre-expand environment variables
