+++ pkg/parser/node/accessor_test.go
@@ -82,26 +82,3 @@ func TestGetSliceOfNonEmptyStrings(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, []string{"command1", "command2"}, slice)
 }
-
-func TestExpandEnvironmentVariablesIsDeterministic(t *testing.T) {
-	env := map[string]string{
-		"C":             "not deterministic",
-		"CI":            "not deterministic",
-		"CIR":           "not deterministic",
-		"CIRR":          "not deterministic",
-		"CIRRU":         "not deterministic",
-		"CIRRUS":        "true",
-		"CIRRUS_BRANCH": "main",
-		"CIRRUS_":       "not deterministic",
-		"CIRRUS_B":      "not deterministic",
-		"CIRRUS_BR":     "not deterministic",
-		"CIRRUS_BRA":    "not deterministic",
-		"CIRRUS_BRAN":   "not deterministic",
-		"CIRRUS_BRANC":  "not deterministic",
-	}
-
-	assert.Equal(t, "main", node.ExpandEnvironmentVariables("$CIRRUS_BRANCH", env))
-
-	assert.Equal(t, "true main", node.ExpandEnvironmentVariables("$CIRRUS $CIRRUS_BRANCH", env))
-	assert.Equal(t, "main true", node.ExpandEnvironmentVariables("$CIRRUS_BRANCH $CIRRUS", env))
-}
